services/user-service/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
grpcio-reflection==1.71.0
pymongo==4.13.2
pydantic[email]==2.7.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.2
dishka==1.6.0
structlog==24.4.0
redis==5.0.8